// src/lib/types/home.ts
export interface Button {
  id: number;
  buttontext: string;
  buttonlink: string;
  buttonclass: string;
}